Update README.md
Added clarification to the core claim of The deVinery Method. 

Sexual fantasies are stories designed by our own minds, to navigate the three paradoxes inherent to desire so we can move from one state of consciousness to another. Sexual fantasies are desire-driven consciousness tech that overcomes our ego resistance (quietens the default Mode Network). The way we each navigate these three paradoxes is unique as a fingerprint but can be precisely mapped. These gives us the underlying engine of each person's erotic mind and can be then used to create infinite variety of real world or role play experiences that satisfy, and change the state of consciousness of that individual. This is the key to consistently being able to lose our ego inhibitions to access sexually aroused flow state, ecstatic states - and sometimes non dual Oneness states. It is the key to sustained sexual desire in longer term relationships.